Do not rely on compiler implementation-dependent #include search strategies.
[Reviewed by sungeun.]

In the runtime we have a few #include files which can be provided by
specific comm or locale model implementations, with generic (usually
empty) last-resort versions of these files available for use when the
implementation does not supply them.  An example of one of these is
chpl-comm-impl.h, which is #included by runtime/include/chpl-comm.h to
provide any special declarations the comm layer implementation needs.  A
comm layer can provide runtime/include/comm/$CHPL_COMM/chpl-comm-impl.h
if it needs it.  But if no special declarations are needed the comm
layer can leave that file nonexistent, and the last-resort version of
it, in runtime/include, will be used.  Because the compiler command line
for runtime builds has -Iruntime/include/comm/$CHPL_COMM before
-Iruntime/include, a layer-provided chpl-comm-impl.h is always found
before the last-resort one.

EXCEPT, it turns out that the technique we have been using to arrange
for the last-resort files to be used when nothing else is available is
compiler dependent, and thus fragile.  The strategy by which a compiler
finds #include files is not specified by the C standard.  It is left to
the implementation.  We have been relying on the gcc strategy.  For an
include file delimited by "", gcc searches first in the directory
containing the file in which the #include appears, and then in any
directories specified by -I options given to the compiler, in left to
right order.  For an include file delimited by <>, it searches in
certain system directories, then in the -I directories.  Unlike the ""
case, in the <> case it does not search in the directory containing the
file in which the #include appears.  So in the Chapel runtime, if a
#include file has a last-resort version, we delimit it with <> when we
#include it.  This avoids the last-resort file incorrectly overriding
the implementation specific one when the #include happens to come from a
file in the same directory the last-resort version resides in.  But this
technique really only works for compilers that behave as gcc does, and
even then it is dependent on a syntactic quirk whose purpose is easy to
miss.  (In fact, I discovered the need to improve this area by wrongly
using "" to delimit an #include that had a last-resort version, and then
struggling for quite some time to figure out why the last-resort version
was overriding the implementation-specific one.)

This commit makes last-resort file handling a little more robust.  It
moves the last-resort files into directories that do not have anything
else in them, adjusts the compiler -I options to add those directories,
and changes the preprocessor #include commands to use the regular ""
delimiters instead of <>.

The file-specific changes are as follows:

runtime/make/Makefile.runtime.include:
  Add runtime/include/comm to the paths searched for #include
  files, after runtime/include/comm/$(CHPL_MAKE_COMM).  Add
  runtime/include/localeModels to the paths searched for #include
  files, after runtime/include/localeModels/$(CHPL_MAKE_LOCALE_MODEL).

runtime/include/chpl-comm.h
  Change #include <> to #include "" for files that might have a
  last-resort version, thus making all #includes of runtime-specific
  files use "", and only #includes of system files use <>.

runtime/include/chpl-comm-heap-macros.h --> runtime/include/comm/
runtime/include/chpl-comm-impl.h --> runtime/include/comm/
runtime/include/chpl-comm-task-decls.h -->runtime/include/comm/
runtime/include/chpl-locale-model.h --> runtime/include/localeModels/
  Move all these last-resort #include files to directories that do
  not contain other files that might #include them.  While doing so,
  add comments clearly identifying them as last-resort files, and
  also, where it's missing, add the usual #ifndef ... #define ... #endif
  protection so that they can effectively only be #included once per
  source file.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21643 3a8e244f-b0f2-452b-bcba-4c88e055c3ca