Adding HDFS support to Chapel
This commit adds HDFS support to Chapel. Information about how to use Chapel with
HDFS as well as how to install Hadoop and get setup can be found in $CHPL_HOME/doc/release/technotes/README.hdfs

The main additions were in src/qio/foreignFS which contains:
  * The HDFS runtime code  (in foreignFS/hdfs)
  * Stubs for error reporting if HDFS is used but not turned on (foreignFS/none)

On top of this, the Makefile system and chplenv were changed slightly to allow the
use of the CHPL_FOREIGN_FS environment variable when making the compiler (just like
regexp support). 

This commit also adds locality aware reading to HDFS and pipes this through to the
module level code (not this is outside the API for the QIO runtime at this point).
If the end user wants to have locality awareness when they are reading a byte range 
from HDFS, they must query this explicitly and then set on which locale the
read is happening themselves based upon this information. 

The API documentation is currently still in progress and will be uploaded sometime
next week. This will also contain future directions and TODO's

[Most of this code reviewed by Michael, and all of it by Tom]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21646 3a8e244f-b0f2-452b-bcba-4c88e055c3ca