Adding notes of a discussion of return intents during the deep dive session 7/30/2013.
Adding it under spec/proposals, as it is mostly a language-related topic.
The discussion involved a little bit of the current implementation status.

The central point is that we agreed that the return mechanism - i.e. by
value vs. by reference - for a non-'var' proc or iter should match the
mechanism used to pass an argument of the same type by blank intent.
See the notes for more.

Updating test/functions/iterators/vass/yield-arrays-var-nonvar.future
and .good accordingly.

TODO: need a similar test/future for a proc (i.e. for returning arrays, rather
than yielding them).

TODO: need a similar test/future for yielding/returning domains.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21653 3a8e244f-b0f2-452b-bcba-4c88e055c3ca