Really fix qthreads tasking layer builds.
[Not reviewed -- simple.]

Yesterday's r21652 fixed the qthreads tasking layer build problems
introduced by r21643, but it also introduced a new problem.  By adding
$CHPL_HOME/runtime/include/comm/$CHPL_MAKE_COMM to the set of paths
searched for #include files, it caused the comm layer's version of
chpl-comm-heap-macros.h to override the private one used by qthreads
tasking, leading to compilation errors.  Fix this by removing that
#include path.

(I failed to notice the problem because I built with comm=none first.
That worked properly and prevented a rebuild of qthreads, which isn't
dependent on the comm layer and so doesn't need to be built separately
for different ones.  But when I built with comm=gasnet first, as last
night's testing did, then the problem occurred.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21655 3a8e244f-b0f2-452b-bcba-4c88e055c3ca