Removed the no-local entry for "counts leaked by domainMaps" from REGRESSIONS.
[reviewed by Sung]

The original regression was due to r21429. However, as I explained on 6/19,
the change in the comm counts was due to modified Chapel sources. It was
not due to any new issues introduced in that commit.

This position was reflected in r21506, which resolved the regression
(by updating .good files).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21656 3a8e244f-b0f2-452b-bcba-4c88e055c3ca