Remove extraneous localeID_to_locale calls for --local compiles
[reviewed by vass, hilde]

With the hierarchical locale merge, calls to localeID_to_locale()
were added to the generated code for a --local compile which had
not been there before.  This had an adverse impact on performance
as seen in the chameneos-redux shootout benchmark, for example.
This commit removes those calls which are side-effect free and
extra overhead for a --local compilation.  This commit squashes
those extra calls for --local compilation and adds a bunch of
comments to document the existing code and note TODO items that
came up during the code review.

It should be noted that as we move to hierarchical locales, 
assuming that on-clauses are used to move computation to sublocales
within a node, we may need to expand the concept of --local to
distinguish between the current "drop all on-clauses" behavior to
one that distinguishes between on- and off-node locality.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21662 3a8e244f-b0f2-452b-bcba-4c88e055c3ca