This is an update to the makefiles that contain Wdelcaration-after-statement.
The original plan was to get rid of this flag as it forces c89 like
declarations and as the name implies you cant have declarations after
statements which is somewhat archaic and not necessary  with c99. 

However, we discovered that MSVC still does not support many features of c99
including mixing declarations and statements. While Chapel doesn't currently
support windows natively it seems premature to get rid of this flag if later
windows support might force you to have to enable it and change code that has
mixed decls. Note that this lack of c99 support is a well known issue with MSVC
and with version 2013 they are supposed to be including more c99 features, but
it's not clear if that will include allowing mixed decls and statements. 

So far as we know portability to MSVC is the only reason to prevent mixing
declarations and statements (and keep this flag around.) It was originally 
included because it was used by the gasnet makefile stubs that we were using,
however getting rid of this warning and mixing declarations and statements 
did not seem to cause any issues with gasnet.

All this commit does is serve as a place to capture why we're leaving the flag
in place and add comments next to the flags pointing to this commit message. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21665 3a8e244f-b0f2-452b-bcba-4c88e055c3ca