Add a framework for inserting per-pass consistency checks. [reviewed by Vass]
This patch adds hooks that can be used to perform consistency checks after each
pass.  This is an enhancement over verify(), since the latter can contain only
checks that are expected to be true across all passes.  

To support adding detailed checks without slowing down a normal test run, this
patch also adds a "--paranoid" flag to the command line (with a corresponding
CHPL_PARANOID environment variable).  New checks (provided as examples; not
intended to represent a complete set) have been placed inside fParanoid
conditionals.

Generally, we expect passes to perform normalizations or optimizations.
Application of a normalization means that additional checks should hold at the
end of that pass.  Application of an optimization means that the same set of
checks should hold at the beginning and end of the pass.  Therefore, we expect
the set of applicable checks to expand as we progress through the compilation
process.

The key construct is the addition of a check_fn field to the PassInfo structure
(Vass's idea).  The RUN() macro fills this in with a pointer the function named
"check_<passname>".  The file checks.cpp provides definitions for these
functions.  These have been populated with an initial set of checks.  It is
anticipated that these checks will be refined and augmented over time.  Writing
and adding checks (especially paranoid checks) is encouraged.

[tested on standard tests]

Details:
    compiler/include/driver.h
    compiler/main/driver.cpp
Add --parnoid flag to command line.

    compiler/include/passes.h
Added a global "parsed" boolean, to indicate when parsing is complete.

    compiler/include/passes.h
    compiler/AST/checkAST.cpp
    compiler/AST/Makefile.share
Added some example checks.
 
    compiler/main/checks.cpp
Added definitions for the per-pass check functions.

    compiler/main/runpasses.cpp
    compiler/main/passlist.h
Added a check_fn field to facilitate calling per-pass check functions.
Moved verify() into check_afterEveryPass().

    compiler/AST/alist.cpp
    compiler/AST/expr.cpp
Added checks that the parentExpr is NULL before insertion.

    compiler/passes/filesToAST.cpp
Added a comment on the block that sets the scoping for SET_LINENO.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21669 3a8e244f-b0f2-452b-bcba-4c88e055c3ca