Changing behavior of 'hdfs' detector script if paths not set/don't exist
In Tim's original commit of this, he made the detection of 'hdfs'
switch to 'none' in the event that things like JAVA_INSTALL and
HADOOP_INSTALL were not set up properly.  In the event that the user
had explicitly requested 'hdfs', this was a little confusing (and
inconsisent with what we do in other scripts), so we switched it to
issue a warning.  In the event that the paths aren't set/available,
the warning will print out and the 'make' step in $CHPL_HOME will
fail (due to missing headers and the like).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21671 3a8e244f-b0f2-452b-bcba-4c88e055c3ca