Improving nbody-blc.chpl and capturing previous version for comparison
In looking at differences between Albert's nbody versions, I noticed
that there's a big jump between iterating over an array of records
via:

        for b2 in B[i+1..] {

and:

        for j in i+1..n-1 {
          const b2 = B[j];

so I made the similar change in mine, to dramatic effect.  Given the
magnitude of difference, it seems likely that this difference is more
than just the overhead of array slicing; or if it is array slicing,
that we should be able to optimize this significantly in the 1D local
case, so I kept the previous version around for historical purposes,
renaming it as nbody-blc-slice.chpl.

I believe that this makes my version the fastest that we have, but
we'll see if tonight's runs confirm that.

Next steps: See whether vectorization comes into play...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21689 3a8e244f-b0f2-452b-bcba-4c88e055c3ca