Fixing mistake in .graph file
I accidentally forgot to update one of the entries in my .graph file,
resulting in warnings/errors last night.  Part of the reason I didn't
test the graphing is that nbody currently takes a long time to run.
While here, I also removed the TIMEOUT on the argument that the cases
we care about most are running in a very reasonable amount of time,
so let's cut down on the time spent waiting for the others to timeout.
Over time, I hope to study the cases timing out and remove them if
they don't seem of sufficient interest.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21691 3a8e244f-b0f2-452b-bcba-4c88e055c3ca