Change to start_test to force the setting of a performance directory when running with -performance and -generate-graphs [reviewed by Sung]
Forces the tester to name the directory where all the performance data will be 
placed.  Additionally fixes a bug this had revealed where graph generation 
tried to access a field that did not exist because not all performance 
tests had been run.

util/start_test:
    checks if CHPL_TEST_PERF_DIR has been set when running with 
-generate-graphs (which also happens when running with -performance)
and if not, aborts start_test

util/test/genGraphs:
    Detects if a file cannot be opened because it does not exist and allows 
the overall program to continue to run, generating a warning to the user 
about this test not being present.  Additionally makes sure that 
nonexistent attributes are not accessed.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21692 3a8e244f-b0f2-452b-bcba-4c88e055c3ca