Bug fix: restores genGraphs CSV creation [Reviewed by: Sung]
The earlier adjustment to genGraphs to allow performance tests that were not
run to be skipped had inadvertently prevented CSV generation for all graphs.
This patch alters genGraphs so that skipping these tests allows CSV generation
to continue.

There are a couple of mods in here to make this possible and to clean things
up.  First, when it is discovered that a file does not exist, we ignore the
I/O warning because it will be taken care of later.  All other exceptions
continue to be raised.  Second, I was erroneously checking if the callee file
had a dfile, rather than the current file it was iterating over.  Realizing
that in the case where the last file in GRAPHFILES had not produced a .dat
file, my current implementation would halt early instead of generating data
for all the .dat files that existed, I changed the return statement to a
deletion of the DatFileClass for a dat file that did not exist.  The final
change occurred in generateData.  Here, with the previous fix, it was trying
to access the DatFileClass of a .dat file that did not exist and hence had
been deleted from the dictionary.  To avoid this error, I had the loop check
that the filename was for a .dat file that existed and if not, continue on to
the next filename. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21699 3a8e244f-b0f2-452b-bcba-4c88e055c3ca