Bringing NCCS qsub options up-to-date [contributed by Saliya Ekanayake]
In the PBS launchers, we've traditionally had special-case code to
handle the wrapper around qsub that NCCS machines like Jaguar/Titan
use to launch jobs.  It seems that the option syntax has chnaged since
the last time we tried it, so Saliya Ekanayake contributed the
following patch to launch-pbs-aprun.c to address it.  I propagated the
similar change to launch-pbs-gasnetrun_ibv.c (though to my knowledge
NCCS' qsub may not be used in an IBV configuration)





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21702 3a8e244f-b0f2-452b-bcba-4c88e055c3ca