Add a hint type in qio_file_s to distinguish owned from unowned file pointers. [reviewed by mferguson@ltsnet.net]
The Problem: Reclamation of the qio_file_t representing stdout caused subsequent
references to stdout to fail.  That is because _qio_file_do_close was calling
fclose() on stdout.

The trigger: Improvements to memory management in the compiler (to appear in a
separate commit) exposed this problem, because qio_file_t objects are now being
reclaimed as the program exits, but additional code attempts to write to stdout
after execution of the Chapel program is complete.

The solution: I added a QIO_HINT_OWNED flag to the hints that can be associated
with a file in a qio_file_t object.  When the file is opened by QIO, this flag
is set to true; when the qio_file_t object is destroyed the file is then
closed.  If instead the file comes from an external source, then it is unowned
and therefore not closed when the qio_file_t object is destroyed.

In particular, if qio_file_init() is called through qio_file_open(),
qio_file_open_usr() or qio_file_open_tmp(), the QIO_HINT_OWNED flag is set
because the qio implementation opened the file internally.  If qio_file_init()
is called directly, the user (or system) is responsible for closing the
associated file.

[tested on examples/, users/ferguson/]

Details:
Index: runtime/include/qio/qio.h
Added the QIO_HINT_OWNED flag to the anonymous qio hints enum.
Enabled some code for debugging channel reference counts that was commented out.

Index: runtime/src/qio/qio.c
In _qio_file_do_close(), close the underlying file implementation only if the
QIO_HINT_OWNED bit is set.
Set this bit if the file implementation is opened by QIO.  Otherwise, leave it unset.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21703 3a8e244f-b0f2-452b-bcba-4c88e055c3ca