Backing out uncommented printing of QIO ref counts
r21703 included the change:

* Enabled some code for debugging channel reference counts that was
  commented out.

But this enabling this seems to have broken the build on Mac OS X in
developer mode because, from what I can tell, a refcount struct type
was being printed out with a %ld format string.  For the time being, I
reverted these to the previous commented out lines to get a clean
build on Mac tonight.  As far as I know they could be uncommented
without adding in the printf of the refcounts, but I didn't want to
investigate that since my goal was to get a clean build in quickly.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21705 3a8e244f-b0f2-452b-bcba-4c88e055c3ca