initial commit of 9-point stencil studies
This is a set of 9-point stencil codes that I've been studying
this evening.

* stencil9.chpl is the basic 9-point stencil for shared-memory
* stencil9-block.chpl is a block-distributed version
* stencil9-explicit.chpl is an SPMD-ized version

The stencil9-explicit.chpl is the interesting one -- the others
are included for the purposes of comparison.  The nice part
about it is the use of an array of domains to store the overlap
panels and the fact that the global index space permits us to
use those panels to refer to both the local and remote sub-blocks.

There's much that could be done to improve the naming of variables and
types, clean up the code, and make it more efficient.  In particular,
I used a lame sync variable idiom in order to get some deterministic
printing for debugging purposes and to avoid races that could be done
more intelligently.  I didn't worry about that tonight due to the
focus on the stencil/fluff communication itself and the hour.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21706 3a8e244f-b0f2-452b-bcba-4c88e055c3ca