Changed start_test for better placement of check for CHPL_TEST_PERF_DIR existence and to give it a default value/generate a warning instead of generating an error when it is not already found. [Reviewed by: Brad]
Changed the start_test behavior when the tester has not set CHPL_TEST_PERF_DIR
from an error to a warning with a default value of $CHPL_HOME/test/perfdat.
Additionally, I moved the check and setting of this environment variable from
within the test for the -generate-graphs flag to within the already present
check for the existence of the environment variable as it seems silly to
separate the two cases and I hadn't noticed the other one before.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21710 3a8e244f-b0f2-452b-bcba-4c88e055c3ca