Prevent incorrect dead-code removal of chpl_here_free() calls.
[Reviewed by hildeth.]

This fixes changes in leaked memory observed after my r21707 commit that
corrected alloc/free mismatches.  It adds the "essential" flag to the
PRIM_HERE_FREE primitive, which is necessary in order to indicate that
chpl_here_free() has side effects and thus should not be removed even
when it appears to be dead code.  (The memory leaks were due to dead
code removal of instances of this primitive.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21730 3a8e244f-b0f2-452b-bcba-4c88e055c3ca