Remove pragmas introduced unnecessarily to "fix" r21707.
[Reviewed by hildeth.]

This undoes part of what I committed in r21727, to fix the compiler
internal errors caused by r21707.  Specifically, it removes the "insert
line file info" pragmas.  It turns out that those are not required.  The
change to the primitive definition to force the presence of line and
file arguments on the calls was all that was needed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21731 3a8e244f-b0f2-452b-bcba-4c88e055c3ca