Use ref and const ref intent for extern procs in IO.chpl
Reviewed by Vass, fixing a problem brought up by Vass.

Replace all uses of the intent 'inout' with 'ref' in IO.chpl.
Also made many of those 'const ref' and tidied up a few things
in the C to be const something* in those cases. Lastly,
removed unused variable in qio.c reported by Rafael Asenjo.

Ref intents are preferable to inout in IO.chpl since most
of these routines are calling C routines which are really
using pass by ref. Using inout would cause a copy to be
made in order to pass it by reference.. which just seems
to add unnecessary work.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21733 3a8e244f-b0f2-452b-bcba-4c88e055c3ca