Enhanced lvalue checking for out/inout/ref/const-ref actuals.
[reviewed by Tom]

In functionResolution.cpp:

* Where it checks that the actuals are lvalues, I added a check
for 'const ref' actuals.

  The check for 'const ref' is a little different. We have discussed
dropping the distinction between lvalues and rvalues, so that
everything can be treated as a (potentially const) lvalue.
Therefore anything could potentially be passed by 'const ref'.
However, today the compiler has this distinction and, furthermore,
does not implement passing of params (e.g. literals) by const ref.
Those cases - but not other constants - will be flagged as an error
for now.

* Refactored those checks (a) to use a switch over intents,
(b) at Tom's request, moved the code that checks the actual
into isLegalLvalueActualArg() and isLegalConstRefActualArg().
These are separate because 'const ref' is different - see above.

* Converted these checks from USR_FATAL to USR_FATAL_CONT,
so multiple errors can be reported in one compiler run.

* Unrelatedly, skip a piece of code if !fn->inTree(),
to avoid useless work.


In symbol.cpp/symbol.h

* Added ArgSymbol::intentDescrString() and char* intentDescrName[IntentTag]
that return a description of the argument's intent.

I made them this way so they can be used in an English sentence,
but feel free to fiddle with that.

* (Could also have const char* intentName[] and ArgSymbol::intentString()
that would generate the actual Chapel intent keyword.)


Test changes:

* Changed a couple of .good files to adjust to the new error message format.

* Added a test for the 'const ref' error message.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21735 3a8e244f-b0f2-452b-bcba-4c88e055c3ca