Fix for the ref intent bug associated with gasnet-fast. [reviewed by diten]
test/users/ferguson/ref-intent-bug.chpl was reporting the error:
  ref-intent-bug.chpl:16: error: unexpected case [parallel.cpp:708]
when tested with CHPL_COMM=gasnet and CHPL_GASNET_SEGMENT=fast.

The problem was most likely due to an argument being passed by reference into
the current function, and the current function passing this along to a
subroutine.  Code in makeHeapAllocations() expected, the value returned through
a ref argument to be established by moving the address of the result into the
ref argument, and not in any other way.

But obviously, passing the ref arg to an arbitrary subroutine (expecting a ref
argument in that position) will also serve to establish the returned value.  So
the cases we are likely to encounter in fact are much more extensive than what
was expected in the existing version of this routine.

Since explicit ref arguments were added to the language relatively recently and
have therefore not been used extensively, it seems reasonable that this case has
remained hidden up to now.

[tested with gasnet-fast: domains/, distributions/, multilocale/, performance; no-local]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21748 3a8e244f-b0f2-452b-bcba-4c88e055c3ca