Update to .good file due to replicate rootLocale change
This one was missed with my replicated rootLocale change, in part
because at the time of the hierarchical locale change, it wasn't
counting comm for the surrounding coforall.  I changed the test in
r21581 (post hloc merge) to be deterministic, but got the comm counts
from that point in time.  With the replicated rootLocale change
(r21725), The extra look ups back to locale 0 were eliminated.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21751 3a8e244f-b0f2-452b-bcba-4c88e055c3ca