Make printchplenv print out short versions of compile,runtime, or launcher dirs
After some discussion on the mailing list, there is a desire to get
new config variables more easily included in the various bin
subdirectories named from the configuration. This patch makes
printchplenv produce such directories, so that there is really
one main place that these configurations are listed. A future patch will
adjust the Makefiles to use the output of this script.

While I was there, I made printchplenv also have the ability to
output a sh or csh script to copy an environment.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21754 3a8e244f-b0f2-452b-bcba-4c88e055c3ca