Removing futures that are now passing
These futures are now passing as a result of the Michael/Tom
fixes from yesterday, so retiring them.

ref-intent-bug-2... is not passing, but is failing due to
a nil dereference issue at execution time, so should
arguably be recategorized.  We've seen other instances of
this recently, some of which seem to be new since 1.7.0,
so should look into it prior to the release to make sure
it doesn't represent a regression.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21758 3a8e244f-b0f2-452b-bcba-4c88e055c3ca