Initial commit of GASNet 1.20.2
This is a commit of the latest GASNet for potential inclusion in the
upcoming release.  This is just a snapshot of the released .tar.gz
along with its unpacked sources.  It doesn't yet apply our standard
modifications, not switch Chapel over to use this version by default.
That will come in subsequent commits.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21760 3a8e244f-b0f2-452b-bcba-4c88e055c3ca