Some changes so it works fine with slurm. One specific line for the University of Malaga computer, Picasso, has been commented, so it can be used as guide for other computers.
It request an exclusive access to the computer nodes, as it is easier to
use in this way. On one hand almost all supercomputers give exclusive
access, so that would have no effect, on the other hand this way it
doesn't need to request an specific number of cores inside the node,
as all are reserved.

Also only one task is requested per computer node, as chapel only needs
one "seed" task that will spawn the needed processes or threads.



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/pario2@21764 3a8e244f-b0f2-452b-bcba-4c88e055c3ca