Reset (most) CyclicDist comm counts tests
We lost track of these for a while, so this commit resets the comm
counts for all the tests except assign.chpl and reduce.chpl.

History for history's sake:

We accidentally stopped running the nightly CyclicDist tests 9/22/12.
We started again 2/6/13.  During that time, the comm counts for all
the tests except assign.chpl and reduce.chpl got much worse.  There is
nothing obvious in the commit logs that would suggest a reason for
this (comm counts for BlockDist did not change).

Both assign.chpl and reduce.chpl started failing 5/30/13 due to r21426
(early outline task functions).  This problem is not fully diagnosed,
but we are aware of it and will return to it at some point.

Then the comm counts for the other tests improved significantly on
6/3/13 with r21457 (shifted base pointer).  In fact, alloc and
alloc_all are doing better than what is in their .good files, and the
others are far less bad than they used to be prior to this commit.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21771 3a8e244f-b0f2-452b-bcba-4c88e055c3ca