Updating nightly.chpbld01 to reflect OS/module upgrade
Now that we've updated this machine to support XC whitebox
testing, updated this script to remove things that I believe
were specific to our XT environment, in part by comparing
to what we do in nightly.chpbld02.

TODO: nightly.chpbld02 is still setting the environment
variables to cray-xt, whereas it should probably be updated
to cray-xe.  I'm going to save that for tomorrow to avoid
breaking too many things in one night.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21774 3a8e244f-b0f2-452b-bcba-4c88e055c3ca