Revert the removal of the the locale.id optimization
[reviewed by David, Greg]

This commit brings back the locale.id optimization that was for some
reason removed in r21480 (hierarchical locales merge).

While I (and others) agree with the comment in there that we shouldn't
do optimizations during parsing, it should not have been removed
without a replacement.

This resolves the remaining comm counts regressions.  In addition, for
some reason alloc_all improved with r21725 (replicated rootLocale), so
update that .good file.  We didn't notice it at the time since it was
failing already.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21776 3a8e244f-b0f2-452b-bcba-4c88e055c3ca