Remove the r21675 stopgap repair.
[Reviewed by hildeth.]

My r21672 commit caused almost 300 failures (not "about 35", as I said
at the time) in baseline testing due to errors like this one:

  arrayassign.chpl:10: error: halt reached - array index out of bounds: (1)

The array index that was out of bounds was derived from a bogus nodeID,
and that in turn was the result of the data object initialization code
being confused by the missing members in the chpl_localeID_t record
type.  In r21675 I provided a stopgap repair for this problem, by adding
the missing members back in.

This commit provides a much more elegant solution, and is due to Tom.
In retrospect it's obvious, but the basic problem is that the compiler
cannot know how to build a copy constructor properly when some or all
of the members of a record type are unknown to it.  The solution is for
the defining entity of such a type (the runtime in this case) to also
provide a copy constructor for it, and to tell the compiler to use that
instead of trying to create one.  That's what we do here.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21778 3a8e244f-b0f2-452b-bcba-4c88e055c3ca