Remove unused CHPL_*_UBAR make variables.
[Reviewed by sungeun.]

One of the features of the runtime build is that you can define a make
variable whose name is of the form <filename_ubar>_CFLAGS, and the build
will add the value of that variable to the compiler command line when it
compiles <filename>.  Here <filename_ubar> is the source filename,
without an extension, and with any characters not legal in make variable
names changed into '_' (underbar) characters, hence the name.  To
support this we have been defining CHPL_MAKE_{TASKS,THREADS,TIMERS}_UBAR
variables, which are the tasking, threading, and timer runtime layer
names, respectively, also with any characters not legal in variable
names changed into '_'.  As it turns out we were only using this for the
Cray-specific tasking and threading layers, and not at all for any timer
layer, and it would actually be easier to understand what was going on
if we just used the layer name itself rather than the CHPL_*_UBAR
variables.  (We used to use this UBAR name for more things, but those
uses have disappeared over time.)  An earlier commit (to the Cray repo)
removed the uses of these *_UBAR variables.  This one removes the
instances where we set them.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21782 3a8e244f-b0f2-452b-bcba-4c88e055c3ca