Feature request future for 'ref' declarations
This feature captures a compelling example for 'ref' declarations,
a feature which we've discussed and stated our intent to support,
but not yet gotten to.  It avoids repeated array accesses (for
clarity and, in a naive implementation, performance reasons) by
using a ref declaration instead.  The alternative that I'm using
in my main implementations is an inlined nested proc -- yuck.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21788 3a8e244f-b0f2-452b-bcba-4c88e055c3ca