This test deadlocks; we'll solve that later; for now make it time out faster.
[Reviewed in discussions with bradcray.]

A recent change (r21725) added a forall in InitPrivateGlobals.chpl, with
the result that in programs running on more than one locale, we need a
minimum of two tasks running simultaneously in order to avoid deadlock.
These two tasks are (1) the one that originally encounters the forall
on locale 0 and (2) a remote task targeted at locale 0 and originating
on some other locale, which performs an atomic update on the endcount
associated with the forall.  With the tasks=fifo this means that we also
need two threads, to host the tasks.  If CHPL_RT_NUM_THREADS_PER_LOCALE
is set to 1 (one), we deadlock because we can't run the task to do the
atomic update on the endcount for the forall.

For now we are leaving this deadlock failure as a reminder to ourselves
to modify the module code so that it works properly with only a single
thread.  But to speed things up in the meantime, here I'm committing a
30-second timeout on this test.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21797 3a8e244f-b0f2-452b-bcba-4c88e055c3ca